feat(modularization): S-11 -- extract settings + auth/security to web/modules/settings.js
Extracted autonomy tab (lines 2131-2220) and settings core with auth/security
(lines 2427-3051) to web/modules/settings.js (~730 lines).

Key seams:
- isSettingsDirty() exported; app.js leave hook uses !isSettingsDirty()
- wireBranchDriftBanner injected via initSettings({wireBranchDriftBanner}) DI,
  called explicitly from app.js at startup (DOMContentLoaded can't fire in ES modules)
- wireAuthBanner()/initAuthBanner() called at module eval time, not in DOMContentLoaded
- body.querySelector('#authCreateBtn') scopes DOM search to authCardBody to avoid
  null reference when body is not yet attached to the main document
- switchPage imported from app-core.js (needed by wireAuthBanner)

Approvals section (lines 2220-2426) left in app.js -- separate page, not a settings tab.

Test update: autonomy-standalone-removal.test.ts APP reads from web/modules/settings.js.

Smoke: 0 JS exceptions, 24 nav items, settings page visible.
